+++ app/src/pages/root/App.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import logo from './logo.svg';
 import './App.css';
 
-import Presentation from './Presentation';
+import Presentation from '../Presentation';
 
 function App() {
   return (
